Fix a runtime error in the "HUH" message.
When you hit enter without giving any command,
the game tries to print the message HUH as a format string...
but it's not a valid format string!

    Fortran runtime error: Missing initial left parenthesis in format

Reuse the existing format string 1001 for this message.

This code may have worked fine on VS Fortran back in the day,
for all I know.