Enable the DEBUG command.
The old code threw a runtime error because of the line
`READ(6,7102) II` — unit 6 is for output, not input. Also,
the password prompt didn't have a newline... which is totally
fine, but means that the prompt doesn't show up if you're
piping output through `/bin/asa`. So let's just add an
inauthentic newline.